requirements.txt
fastapi
uvicorn
pydantic[email]
supabase
requests
python-dotenv